// utils/types.ts
export type UserRole = 'administrator' | 'operator_unit' | 'validator' | 'verifikator';

export type ActiveMenu =
  | 'dashboard'
  | 'master'
  | 'master_renstra'
  | 'pengaturan_kinerja'
  | 'input_kinerja'
  | 'capaian_bulan'
  | 'capaian_triwulan'
  | 'lhe';

export interface User {
  id: string;
  name: string;
  nip: string;
  email: string;
  role: UserRole;
  roleTitle: string;
  opdId: string;
  opdName: string;
  avatarUrl?: string;
}

export interface OPD {
  id: string;
  kode: string;
  nama: string;
  kepala: string;
  nipKepala: string;
  kategori: string;
  email: string;
  telepon: string;
}